lambda/lemputer-auth/requirements.txt
PyJWT==1.4.2
boto3==1.4.3
